requirements2.txt
altair
pandas
scikit-learn
streamlit
seaborn
plotly.express
